done api gateway jwt

+++ borrowingservice/src/main/java/com/tanthanh/borrowingservice/command/api/events/BorrowingEventsHandler.java
@@ -7,12 +7,17 @@ import org.springframework.stereotype.Component;
 
 import com.tanthanh.borrowingservice.command.api.data.BorrowRepository;
 import com.tanthanh.borrowingservice.command.api.data.Borrowing;
+import com.tanthanh.borrowingservice.command.api.model.Message;
+import com.tanthanh.borrowingservice.command.api.service.IBorrowService;
 
 @Component
 public class BorrowingEventsHandler {
 
 	@Autowired
 	private BorrowRepository borrowRepository;
+	
+	@Autowired
+	private IBorrowService borrowService;
 	
 	@EventHandler
 	public void on(BorrowCreatedEvent event) {
@@ -24,12 +29,21 @@ public class BorrowingEventsHandler {
 	}
 	@EventHandler
 	public void on(BorrowDeletedEvent event) {
-		borrowRepository.deleteById(event.getId());
+		if(borrowRepository.findById(event.getId()).isPresent()) {
+			borrowRepository.deleteById(event.getId());
+		}
+		else return;
+		
 	}
 	@EventHandler
 	public void on(BorrowingUpdateBookReturnEvent event) {
-		Borrowing model = borrowRepository.getById(event.getId());
+		Borrowing model = borrowRepository.findByEmployeeIdAndBookIdAndReturnDateIsNull(event.getEmployee(), event.getBookId());
 		model.setReturnDate(event.getReturnDate());
 		borrowRepository.save(model);
+	}
+	@EventHandler
+	public void on(BorrowSendMessageEvent event) {
+		Message message = new Message(event.getEmployeeId(), event.getMessage());
+		borrowService.sendMessage(message);
 	}
 }
